Cambio de nombre a estandarización, se elimino el modulo data_loading y su funcion quedo dentro del modulo estandarizacion.py, los otros modulos los deje comentados mientras trabajo en este dentro del archivo analisis para evitar errores mientras se organiza
